+++ apibackuper/__init__.py
@@ -3,6 +3,6 @@ apibackuper: a command-line tool and python library for API backuping
 
 """
 
-__version__ = '1.0.1'
+__version__ = '1.0.3'
 __author__ = 'Ivan Begtin'
 __licence__ = 'MIT'