+++ apibackuper/constants.py
@@ -1,2 +1,3 @@
 DEFAULT_OPTIONS = {'initialized' : False}
 DEFAULT_DELAY = 0.5
+FIELD_SPLITTER = '.'