+++ HISTORY.rst
@@ -3,7 +3,20 @@
 History
 =======
 
+1.0.3 (2020-10-28)
+------------------
+* Added several new options
+* Added aria2 download support for files downloading
+
+
+1.0.2 (2020-09-20)
+------------------
+* Using permanent storage dir "storage" instead of temporary "temp" dir
+* Added logic to do requests to get addition info on retrieved objects, command "follow"
+* Added logic to retrieve files linked with retrieved objects, command "getfiles"
 
 1.0.1 (2020-08-14)
 ------------------
 * First public release on PyPI and updated github code
+
+